rename(search): rust_daemon.py -> daemon.py; RustSearchDaemon -> SearchDaemon
The Python-side daemon file is again `daemon.py` — it's the only
daemon that ships now.  Class renamed to match.

Also drops:
  rust/services/search-plugin/PARITY_ROADMAP.md — per task #41, the
    roadmap goes away with the migration it tracked (see
    feedback_docs_principles.md rule 'No migration roadmap docs').
  docs/operations/search-mutation-parking.md   — describes the
    Python daemon's parked mutation queue + /parked/skip-to admin
    endpoints, which no longer exist.
